Add missing android package in test module
