+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/helpers/QueryChannelsPostponeHelperTests.kt
@@ -9,7 +9,7 @@ import io.getstream.chat.android.client.Mother
 import io.getstream.chat.android.client.api.ChatApi
 import io.getstream.chat.android.client.clientstate.ClientState
 import io.getstream.chat.android.client.clientstate.ClientStateService
-import io.getstream.chat.test.asCall
+import io.getstream.chat.android.test.asCall
 import org.amshove.kluent.When
 import org.amshove.kluent.calling
 import org.amshove.kluent.shouldBeEqualTo
