+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/BaseDomainTest.kt
@@ -35,8 +35,8 @@ import io.getstream.chat.android.livedata.utils.RetryPolicy
 import io.getstream.chat.android.livedata.utils.TestDataHelper
 import io.getstream.chat.android.livedata.utils.TestLoggerHandler
 import io.getstream.chat.android.livedata.utils.waitForSetUser
-import io.getstream.chat.test.TestCall
-import io.getstream.chat.test.TestCoroutineRule
+import io.getstream.chat.android.test.TestCall
+import io.getstream.chat.android.test.TestCoroutineRule
 import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.When
 import org.amshove.kluent.calling
