+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/BaseDomainTest2.kt
@@ -29,8 +29,8 @@ import io.getstream.chat.android.livedata.controller.QueryChannelsSpec
 import io.getstream.chat.android.livedata.utils.EventObserver
 import io.getstream.chat.android.livedata.utils.RetryPolicy
 import io.getstream.chat.android.livedata.utils.TestDataHelper
-import io.getstream.chat.test.TestCall
-import io.getstream.chat.test.TestCoroutineRule
+import io.getstream.chat.android.test.TestCall
+import io.getstream.chat.android.test.TestCoroutineRule
 import kotlinx.coroutines.asExecutor
 import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.When
