+++ stream-chat-android-test/src/main/java/io/getstream/chat/android/test/CallTestUtils.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.test
+package io.getstream.chat.android.test
 
 import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.client.utils.Result
